Fix: 에러나는 부분 수정

See merge request s09-fintech-finance-sub2/S09P22D209!20

+++ Idontcare/src/main/java/d209/Idontcare/relationship/repository/RelationshipRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 public interface RelationshipRepository extends JpaRepository<Relationship, Long> {
 
   @Query("select count(r.relationshipId) > 0 from Relationship r where r.parent.userId = :parentUserId and r.child = :childUserId")
-  Page<Integer> existsByParentAndChild(@Param("parentUserId") Long parentUserId, @Param("childUserId") Long childUserId, Pageable pageable);
+  Page<Long> existsByParentAndChild(@Param("parentUserId") Long parentUserId, @Param("childUserId") Long childUserId, Pageable pageable);
   
   @Query("select r.relationshipId as relationshipId, r.child.userId as userId, r.child.name as userName, r.createdAt as createdAt" +
       " from Relationship r where r.parent.userId = :parentUserId")
